Move configureWorkspace to src
