+++ src/configureWorkspace/configure.ts
@@ -10,11 +10,11 @@ import * as path from "path";
 import * as vscode from "vscode";
 import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import * as xml2js from 'xml2js';
-import { quickPickWorkspaceFolder } from '../commands/utils/quickPickWorkspaceFolder';
-import { ext } from '../src/extensionVariables';
-import { globAsync } from '../src/utils/async';
-import { extractRegExGroups } from '../src/utils/extractRegExGroups';
-import { Platform, PlatformOS } from '../src/utils/platform';
+import { quickPickWorkspaceFolder } from '../../commands/utils/quickPickWorkspaceFolder';
+import { ext } from '../extensionVariables';
+import { globAsync } from '../utils/async';
+import { extractRegExGroups } from '../utils/extractRegExGroups';
+import { Platform, PlatformOS } from '../utils/platform';
 import { promptForPort, quickPickOS, quickPickPlatform } from './config-utils';
 import { configureCpp } from './configure_cpp';
 import { configureAspDotNetCore, configureDotNetCoreConsole } from './configure_dotnetcore';
