+++ src/configureWorkspace/configure_cpp.ts
