+++ src/extension.ts
@@ -39,8 +39,6 @@ import { systemPrune } from '../commands/system-prune';
 import { tagImage } from '../commands/tag-image';
 import { docker } from '../commands/utils/docker-endpoint';
 import { DefaultTerminalProvider } from '../commands/utils/TerminalProvider';
-import { DockerDebugConfigProvider } from '../configureWorkspace/configDebugProvider';
-import { configure, configureApi } from '../configureWorkspace/configure';
 import { AzureAccountWrapper } from '../explorer/deploy/azureAccountWrapper';
 import { WebAppCreator } from '../explorer/deploy/webAppCreator';
 import { DockerExplorerProvider } from '../explorer/dockerExplorerProvider';
@@ -50,6 +48,8 @@ import { DockerHubImageTagNode } from '../explorer/models/dockerHubNodes';
 import { NodeBase } from '../explorer/models/nodeBase';
 import { browseAzurePortal } from '../explorer/utils/browseAzurePortal';
 import { browseDockerHub, dockerHubLogin, dockerHubLogout } from '../explorer/utils/dockerHubUtils';
+import { DockerDebugConfigProvider } from './configureWorkspace/configDebugProvider';
+import { configure, configureApi } from './configureWorkspace/configure';
 import { COMPOSE_FILE_GLOB_PATTERN, configPrefix, configurationKeys } from './constants';
 import { registerDebugConfigurationProvider } from './debugging/coreclr/registerDebugger';
 import { DockerComposeCompletionItemProvider } from './dockerCompose/dockerComposeCompletionItemProvider';
